+++ formtools/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '2.1'
+__version__ = '2.2'
 
 default_app_config = 'formtools.apps.FormToolsConfig'