+++ docs/changelog.rst
@@ -12,6 +12,8 @@ This page details the changes in the various ``django-formtools`` releases.
 
 - Added support for Django 2.1 and 2.2, and Python 3.7.
 
+- Updated translations from Transifex
+
 2.1 (2017-10-04)
 ----------------